added refresh all added deleting player achievement after completion and creation of completedAchievment only sum types can be assigned to acheviements added cascade deletion verified privacy middleware added getPointsForGameAcc suspend progress on inactive achievements

+++ src/models/achievementDefinition.js
@@ -34,14 +34,61 @@ const AchievementDefinitionSchema = new Schema({
     }
     //add per categoryMetrics 
 });
-//Auto-delete expired achievements Definitions and player achievements before any .find() call
+// Auto-delete expired achievements before any .find() call
 AchievementDefinitionSchema.pre(/^find/, async function (next) {
     try {
-        await this.model.deleteMany({ expirationDate: { $lt: new Date() } });
-        await playerAchievement.deleteMany({ achievementDefinition: { $in: this._id } });
+        const PlayerAchievement = mongoose.model('PlayerAchievement');
+        const expiredAchievements = await this.model.find({ expirationDate: { $lt: new Date() } });
+        const expiredIds = expiredAchievements.map(a => a._id);
+        
+        if (expiredIds.length > 0) {
+            await PlayerAchievement.deleteMany({ achievementDefinition: { $in: expiredIds } });
+            await this.model.deleteMany({ expirationDate: { $lt: new Date() } });
+        }
+        
         next();
     } catch (err) {
         console.error("Error cleaning expired achievements:", err);
+        next(err);
+    }
+});
+
+// Cascade delete PlayerAchievements - covers doc.deleteOne()
+AchievementDefinitionSchema.pre('deleteOne', { document: true, query: false }, async function (next) {
+    try {
+        const PlayerAchievement = mongoose.model('PlayerAchievement');
+        await PlayerAchievement.deleteMany({ achievementDefinition: this._id });
+        next();
+    } catch (err) {
+        next(err);
+    }
+});
+
+// Cascade delete PlayerAchievements - covers findByIdAndDelete()
+AchievementDefinitionSchema.pre('findOneAndDelete', async function (next) {
+    try {
+        const PlayerAchievement = mongoose.model('PlayerAchievement');
+        const doc = await this.model.findOne(this.getFilter());
+        if (doc) {
+            await PlayerAchievement.deleteMany({ achievementDefinition: doc._id });
+        }
+        next();
+    } catch (err) {
+        next(err);
+    }
+});
+
+// Cascade delete PlayerAchievements - covers deleteMany()
+AchievementDefinitionSchema.pre('deleteMany', async function (next) {
+    try {
+        const PlayerAchievement = mongoose.model('PlayerAchievement');
+        const docs = await this.model.find(this.getFilter());
+        if (docs.length > 0) {
+            const ids = docs.map(d => d._id);
+            await PlayerAchievement.deleteMany({ achievementDefinition: { $in: ids } });
+        }
+        next();
+    } catch (err) {
         next(err);
     }
 });
